# Fix performance problem for `Tables.Columns(::KeyedArray)`