There's a type stability issue which occurs in `Tables.columns` when the type of the keys are different on different axes.

I demonstrate that the problem exists, and that this PR fixes it. Not sure what kind of regression test would be optimal, but happy to add one / some if there's an obvious place to put them.


# MWE
```julia
using AxisKeys, Tables

D1 = 2500;
D2 = 2500;
data = randn(D1, D2);
k1 = collect(1:D1);

# Choose between these two to swap between having the same and different key types.
k2 = Int32.(collect(1:D2));
# k2 = collect(1:D2);

X = KeyedArray(data; k1, k2);
@benchmark Tables.columns($X)
```
on master then output of benchmarking is
```julia
BenchmarkTools.Trial: 8 samples with 1 evaluation.
 Range (min … max):  654.951 ms … 766.341 ms  ┊ GC (min … max):  8.84% … 17.49%
 Time  (median):     696.806 ms               ┊ GC (median):    12.74%
 Time  (mean ± σ):   704.700 ms ±  42.018 ms  ┊ GC (mean ± σ):  13.40% ±  4.94%

  █             ▁      ▁  ▁                    ▁     ▁        ▁  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁█▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁█▁▁▁▁▁▁▁▁█ ▁
  655 ms           Histogram: frequency by time          766 ms <

 Memory estimate: 604.74 MiB, allocs estimate: 22445012.
```

On this branch it's
```julia
BenchmarkTools.Trial: 252 samples with 1 evaluation.
 Range (min … max):  18.942 ms … 24.603 ms  ┊ GC (min … max): 0.00% … 21.99%
 Time  (median):     19.035 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.889 ms ±  1.933 ms  ┊ GC (mean ± σ):  4.34% ±  7.99%

  █▇                                                        ▂  
  ███▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄▁▁▇█▄▇██ ▅
  18.9 ms      Histogram: log(frequency) by time      24.4 ms <

 Memory estimate: 71.53 MiB, allocs estimate: 10.
```
By changing the value of `k2` to comprise `Int64`s, you can verify that this doesn't cause a performance regression for the situation in which the keys are the same type, in which case there isn't presently a type-instability.