@rofinn 2022-08-10
This really kinda feels like a hack around a limitation with Julia? Maybe we should raise this as example against the core language?