@rofinn 2022-08-10
Okay, I realize it's a bit more work, but would you mind extracting this logic out into a separate internal function to:

1. Add a type stability unit test
2. Use as a MWE
3. Make it easier to just delete / rewrite if the stability issue is fixed in the future