@willtebbutt 2022-08-10
Very happy to do that. Could you suggest where I put the type-stability test? I'm not sure what the appropriate location is