@willtebbutt 2022-08-10
Okay. I've factored out the code. The following should suffice as a unit test:
```julia
@inferred AxisKeys._get_keys_columns((randn(5), randn(Float32, 4)))
```
but I'm unsure what the correct location for it is.