@rofinn 2022-08-10
I think at the end of the Tables.jl tests should be fine. Please and thank you!

https://github.com/mcabbott/AxisKeys.jl/blob/master/test/_packages.jl#L106